% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodl.R
\name{.get_dm}
\alias{.get_dm}
\title{Get pivot_column key data frame}
\usage{
.get_dm(df, pivot_column, pivot_value)
}
\arguments{
\item{df}{A long data frame}

\item{pivot_column}{The column name on which the pivot will occur}

\item{pivot_value}{The column name of the values to be pivoted}
}
\description{
Get pivot_column key data frame
}
\keyword{internal}
