% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodl.R
\name{.get_key}
\alias{.get_key}
\title{Get pivot_column key data frame}
\usage{
.get_key(tmp, pivot_column)
}
\arguments{
\item{tmp}{A long data frame}

\item{pivot_column}{The column name on which the pivot will occur}
}
\description{
Get pivot_column key data frame
}
\keyword{internal}
