% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodl.R
\name{key_value_table}
\alias{key_value_table}
\title{Generate a key value table with unique key for a set of text}
\usage{
key_value_table(text)
}
\arguments{
\item{text}{The text to abbreviate and create a key value table for}
}
\description{
Given a vector of characters, this will return a data frame of
a unique \code{key} column (where possible of 3 characters) and \code{value}
column listing the unique elements of the original \code{text}.
}
\examples{
data(wb)
key_value_table(wb$indicator)
}
