% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodl.R
\name{make_key_value}
\alias{make_key_value}
\title{Generate a key value table with unique key for a set of text}
\usage{
make_key_value(text)
}
\arguments{
\item{text}{The text to abbreviate and create a key value table for}
}
\value{
df A \verb{Key Value} table
}
\description{
Given a vector of characters, this will return a data frame of
a unique \code{key} column (of, where possible, 3 characters) and \code{value}
column listing the unique elements of the original \code{text}.
}
\examples{
data(wb)
make_key_value(wb$indicator)
}
