% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\docType{data}
\name{wb}
\alias{wb}
\title{Dummy Long Tidy Data}
\format{
A data frame with 975 rows and 4 variables
\itemize{
\item \strong{iso3c:} isocode of a country.
\item \strong{indicator:} World Bank indicator.
\item \strong{year:} Year of observation
\item \strong{value:} Value of observation
}
}
\usage{
data(wb)
}
\description{
A dataset from the World Bank of a dummy data. The variables are as follows:
}
\keyword{datasets}
