% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodl.R
\name{tidymodl}
\alias{tidymodl}
\title{Creates a model matrix style R6 class for modelling with long tidy data}
\description{
Creates a model matrix style R6 class for modelling with long tidy data

Creates a model matrix style R6 class for modelling with long tidy data
}
\details{
Creates a model matrix style R6 class for modelling with long tidy data
}
\examples{
data(wb)
mdl <- tidymodl$new(wb,
                    pivot_column = "indicator",
                    pivot_value = "value")
### Use mdldata for modelling
fit = lm(data = mdl$mdldata, gni ~ gcu + ppt)

### Can be used to add a yhat value for processed data

### This is useful for imputation purposes as below

### NOT RUN
# Use for xgboost imputation
# library(mixgb)
# imp = mixgb(mdl$mdldata, save.models = T)
# tmp = mdl$reconstitute(imp$imputed.data[[1]])

### NOT RUN
# Use for mice imputation
# library(mice)
# imp = mice(as.data.frame(scale(mdl$matrix)), print = FALSE)
# tmp = mdl$reconstitute(complete(imp))

### In this example we will just use dummy new data
nc = ncol(mdl$mdldata)
nr = nrow(mdl$mdldata)
dm = nc*nr
dummy = matrix(runif(dm),
        ncol = nc) |>
        data.frame()
names(dummy) = names(mdl$mdldata)
tmp = mdl$reconstitute(dummy)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{(\code{data.frame()})\cr
The original tidy long data frame}

\item{\code{mdldata}}{(\code{data.frame()})\cr
The model matrix version of the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tidymodl-new}{\code{tidymodl$new()}}
\item \href{#method-tidymodl-reconstitute}{\code{tidymodl$reconstitute()}}
\item \href{#method-tidymodl-print}{\code{tidymodl$print()}}
\item \href{#method-tidymodl-clone}{\code{tidymodl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tidymodl-new"></a>}}
\if{latex}{\out{\hypertarget{method-tidymodl-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.


Create a new tidymodl object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tidymodl$new(df, pivot_column, pivot_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A tidy long data frame}

\item{\code{pivot_column}}{The column name on which the pivot will occur}

\item{\code{pivot_value}}{The column name of the values to be pivotted}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{tidymodl} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tidymodl-reconstitute"></a>}}
\if{latex}{\out{\hypertarget{method-tidymodl-reconstitute}{}}}
\subsection{Method \code{reconstitute()}}{
Adds a results matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tidymodl$reconstitute(m = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{A results matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tidymodl-print"></a>}}
\if{latex}{\out{\hypertarget{method-tidymodl-print}{}}}
\subsection{Method \code{print()}}{
Prints the key and the head matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tidymodl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tidymodl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tidymodl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tidymodl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
